+++ Source/Core/Visualization/Event/KeyEvent.h
@@ -1,19 +1,10 @@
 /****************************************************************************/
 /**
- *  @file KeyEvent.h
- */
-/*----------------------------------------------------------------------------
- *
- *  Copyright (c) Visualization Laboratory, Kyoto University.
- *  All rights reserved.
- *  See http://www.viz.media.kyoto-u.ac.jp/kvs/copyright/ for details.
- *
- *  $Id: KeyEvent.h 1325 2012-10-04 10:34:52Z naohisa.sakamoto@gmail.com $
+ *  @file   KeyEvent.h
+ *  @author Naohisa Sakamoto
  */
 /****************************************************************************/
-#ifndef KVS__KEY_EVENT_H_INCLUDE
-#define KVS__KEY_EVENT_H_INCLUDE
-
+#pragma once
 #include <kvs/EventBase>
 #include <kvs/Key>
 
@@ -28,14 +19,13 @@ namespace kvs
 /*==========================================================================*/
 class KeyEvent : public kvs::EventBase
 {
-protected:
-
+private:
     int m_key; ///< key code
     int m_x; ///< mouse cursol position x
     int m_y; ///< mouse cursol position y
+    int m_modifiers; ///< modifier keys
 
 public:
-
     KeyEvent();
     KeyEvent( const KeyEvent& event );
     KeyEvent( int key, int x, int y );
@@ -44,12 +34,12 @@ public:
     int key() const;
     int x() const;
     int y() const;
+    int modifiers() const;
     int type() const;
 
     void setKey( int key );
     void setPosition( int x, int y );
+    void setModifiers( int modifiers );
 };
 
 } // end of namespace kvs
-
-#endif // KVS_CORE_KEY_EVENT_H_INCLUDE